create navbar
