+++ task-roulette/src/App.js
@@ -3,6 +3,7 @@ import Home from './components/home';
 import Mainpage from './components/mainpage';
 import Dailytask from './components/dailytask';
 import Newtask from './components/newtask';
+import Navbar from './components/navbar';
 import {createBrowserRouter, RouterProvider} from 'react-router-dom';
 
 //What's the difference between BrowserRouter and CreateBrowswerRouter?
@@ -27,10 +28,15 @@ const router = createBrowserRouter([
 ])
 function App() {
   return (
-    <RouterProvider
-    router={router}
-    fallbackElement={<Home />}
-    />
+    <createBrowserRouter>
+      <Navbar/>
+      <RouterProvider
+        router={router}
+        fallbackElement={<Home />}
+      />
+      
+    </createBrowserRouter>
+    
   );
 }
