+++ task-roulette/src/styles/navbar.css
@@ -1,0 +1,22 @@
+.flex-container-main {
+    display: flex;
+    flex-direction: column;
+}
+
+.flex-container-btns {
+    display: flex;
+    flex-direction: row;
+    text-align: center;
+    justify-content: center;
+    margin-top: 1em;
+    margin-bottom: 1em;
+}
+
+.iconImg {
+    width: 100px; 
+    margin: 0em 3.5em 0em 3.5em;
+}
+
+.button-title {
+    font-size: 25px;
+}
